test(cli): consolidate cloud run coverage
